requirements.txt
python-telegram-bot[webhooks]==21.9
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
APScheduler==3.10.4
dateparser==1.2.0
pytz==2024.2
